added loading to general settings
